Exercises 4.15 - 4.23

+++ part4/controllers/blogs.js
@@ -1,6 +1,6 @@
 const blogsRouter = require("express").Router();
 const Blog = require("../models/blog");
-const User = require("../models/user");
+const middleware = require("../utils/middleware");
 
 blogsRouter.get("/", async (request, response) => {
   /*Blog.find({}).then((blogs) => {
@@ -19,31 +19,50 @@ blogsRouter.get("/:id", async (request, response, next) => {
   }
 });
 
-blogsRouter.post("/", async (request, response, next) => {
-  const body = request.body;
-  const user = await User.findById(body.userId);
-  if (!body.title || !body.url) {
-    response.status(400).end();
-  } else {
-    const blog = new Blog({
-      title: body.title,
-      author: body.author,
-      url: body.url,
-      likes: body.likes ? body.likes : 0,
-      user: user.id,
-    });
+blogsRouter.post(
+  "/",
+  middleware.userExtractor,
+  async (request, response, next) => {
+    const body = request.body;
 
-    const savedBlog = await blog.save();
-    user.blogs = user.blogs.concat(savedBlog._id);
-    await user.save();
-    response.status(201).json(savedBlog);
+    const user = request.user;
+    if (!user || !body.title || !body.url) {
+      response.status(400).end();
+    } else {
+      const blog = new Blog({
+        title: body.title,
+        author: body.author,
+        url: body.url,
+        likes: body.likes ? body.likes : 0,
+        user: user.id,
+      });
+
+      const savedBlog = await blog.save();
+      user.blogs = user.blogs.concat(savedBlog._id);
+      await user.save();
+      response.status(201).json(savedBlog);
+    }
   }
-});
+);
 
-blogsRouter.delete("/:id", async (request, response, next) => {
-  await Blog.findByIdAndRemove(request.params.id);
-  response.status(204).end();
-});
+blogsRouter.delete(
+  "/:id",
+  middleware.userExtractor,
+  async (request, response, next) => {
+    /* const decodedToken = jwt.verify(request.token, process.env.SECRET);
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: "token invalid" });
+  } */
+    const blog = await Blog.findById(request.params.id);
+
+    const user = request.user;
+
+    if (user && blog.user.toString() === user.id.toString()) {
+      await Blog.findByIdAndRemove(request.params.id);
+      response.status(204).end();
+    } else return response.status(401).json({ error: "invalid user" });
+  }
+);
 
 blogsRouter.put("/:id", async (request, response, next) => {
   const body = request.body;
